requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
httpx>=0.25.0
requests>=2.31.0
packaging>=23.0
tkinterdnd2>=0.4.0
